backend/requirements.txt
asgiref==3.8.1
boto3==1.36.23
botocore==1.36.23
Django==5.1.6
django-cors-headers==4.7.0
django-froala-editor==4.5.0
django-storages==1.14.5
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
jmespath==1.0.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
